Switch from pointers to steps, add tests, add CI.
Switch from using pointers, which rely on JSON pointers (which aren't as
precise as we would like), to using steps, a more precise but
non-standard way to traverse a complex data type. This matches the
upstream we're shamelessly "inspired" by. It also has the nice property
of drastically reducing the number of dependencies we have.

Add tests. To give ourselves the Go API we want but also the JSON
encoding we want, we need to do some clever work in the UnmarshalJSON
and MarshalJSON functions of our Steps type. Implement tests to make
sure we're not being too clever for our own good.

Add CI. Now that we have tests, set up CI for linting, running tests,
and running dependabot.